# website/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indexarr",
    "version": "0.1.0",
    "description": "Decentralized torrent indexing with DHT crawling, content classification, gossip-based P2P sync, and contributor identity.\n\n## Authentication\n\nMost endpoints require no authentication. Admin endpoints (API key management, nuke review, debug logging) require an API key passed via the `X-Api-Key` header. The Torznab endpoint supports an optional `apikey` query parameter.\n\n## Base URL\n\nAll REST endpoints are prefixed with `/api/v1/`. The Torznab endpoint is at `/api/torznab`. GraphQL is at `/graphql`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/recent": {
      "get": {
        "tags": ["Search"],
        "summary": "Recently resolved torrents",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 100 } }
        ],
        "responses": { "200": { "description": "List of recently resolved torrents" } }
      }
    },
    "/api/v1/announced/recent": {
      "get": {
        "tags": ["Search"],
        "summary": "Recently announced torrents",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 100 } }
        ],
        "responses": { "200": { "description": "List of recently announced torrents" } }
      }
    },
...
    "/api/v1/sync/delta/{content_hash}": {
      "get": {
        "tags": ["Sync"],
        "summary": "Download delta",
        "description": "Download a delta file by its content hash. Returns gzip-compressed NDJSON.",
        "parameters": [
          { "name": "content_hash", "in": "path", "required": true, "schema": { "type": "string" }, "description": "SHA-256 content hash of the delta" }
        ],
        "responses": {
          "200": { "description": "Delta file (gzip or NDJSON)" },
          "404": { "description": "Delta not found" }
        }
      }
    },
...
  }
}